Continue to break out components
